+++ xcube/cli/gen2.py
@@ -123,6 +123,7 @@ def gen2(request_path: str,
     from xcube.core.gen2 import CubeGenerator
     from xcube.core.gen2 import CubeGeneratorError
     from xcube.core.gen2 import CubeGeneratorRequest
+    from xcube.util.versions import get_xcube_versions
 
     verbosity = len(verbose) if verbose else 0
 
@@ -149,10 +150,17 @@ def gen2(request_path: str,
                       message=f'{error}',
                       traceback=traceback.format_tb(error.__traceback__))
         if isinstance(error, CubeGeneratorError):
-            if error.remote_output:
+            if error.status_code is not None:
+                result.update(status_code=error.status_code)
+            if error.remote_output is not None:
                 result.update(remote_output=error.remote_output)
-            if error.remote_traceback:
+            if error.remote_traceback is not None:
                 result.update(remote_traceback=error.remote_traceback)
+        if 'status_code' not in result:
+            result['status_code'] = 500
+
+    if result.get('status') == 'error':
+        result['versions'] = get_xcube_versions()
 
     if output_file is not None:
         with open(output_file, 'w') as fp: